Profile Editor populating Spinners and collecting data
The Profile Editor class seems to be working properly. It successfully populates all of the Spinners and successfully collects and records all the data when the submit button is clicked. Still need to handle the picture button being clicked, and need to handle passing all of the data to the MainActivity to create an actual profile and store the data into a file.