EditProfileFragment done, unless other bugs found or features needed.
The EditProfileFragment now takes a picture and transmit data to the main activity to create a UserProfile object ands ave the data to a file when the submit button is clicked. If you do this, then run the app again, it will load nothing, because it reads that there's already a file, so it doesn't open the edit profile fragment.