Profile upload working. Profile download in place
The profile database files are successfully uploading to AWS. There is code in place to download the database files, but it is never called because it's never necesary to do so.

I also made a change in the MainActivity that would update the number of users in the database. There had been some problems where the value was not updated, and after a new user was created they could therefore not click on their title picture to view their profile.